xml 수정 완료

+++ src/main/java/site/metacoding/miniproject2/service/CompanysService.java
@@ -23,19 +23,20 @@ public class CompanysService {
     }
 
     public void Companyupdate(Integer id, CompanysUpdateReqDto companysUpdateReqDto) {
-        // 1. 영속화
         CompanysTitleReqDto companysTitleReqDtoPS = companysDao.findByIdCompanyId(id);
 
         if (companysTitleReqDtoPS == null) {
-            throw new RuntimeException(id + "의 게시글을 찾을 수 없습니다.");
+            throw new RuntimeException(id + "의 회사정보를 찾을 수 없습니다.");
         }
-        companysTitleReqDtoPS.updateCompanys(companysUpdateReqDto);// 변경
+        companysDao.updateCompanys(companysUpdateReqDto);// 변경
         companysDao.updateChangeCompanys(companysTitleReqDtoPS);// 수행
-
     }
 
     public void deleteCompanys(Integer id) {
-        companysDao.deleteCompanys(id);
+        CompanysTitleReqDto companysTitleReqDtoPS = companysDao.findByIdCompanyId(id);
+        if (companysTitleReqDtoPS == null) {
+            companysDao.deleteCompanys(id);
+        }
     }
 
     public CompanyDetailWithWantedsListRespDto findByIdToDetailWithWantedsList(Integer id) {
@@ -43,8 +44,7 @@ public class CompanysService {
             return null;
         CompanyDetailWithWantedsListRespDto companyDetailWithWantedsListDtoPS = new CompanyDetailWithWantedsListRespDto();
         companyDetailWithWantedsListDtoPS.setCompanyDetailRespDto(findByIdToDetail(id));
-        companyDetailWithWantedsListDtoPS.setWantedsListDtos(wantedsService.findByIdCompanyId(id));
-        // 이부분 물어보기
+        companyDetailWithWantedsListDtoPS.setWantedsListRespDtos(wantedsService.findByIdCompanyId(id));
         return companyDetailWithWantedsListDtoPS;
     }
